requirements.txt
asgiref==3.9.1
beautifulsoup4==4.14.3
black==25.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
deep-translator==1.11.4
Django==5.2.4
django-appconf==1.1.0
django-modeltranslation==0.19.17
djhtml==3.0.9
djlint==1.36.4
EditorConfig==0.17.1
idna==3.11
jsbeautifier==1.15.4
json5==0.12.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.1
PyYAML==6.0.2
rcssmin==1.1.2
regex==2025.7.34
requests==2.32.5
rjsmin==1.2.2
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
